pyproject.toml
[project]
name = "kollabium-notification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.21.0",
    "asyncio>=4.0.0",
    "faststream[nats]>=0.5.48",
    "icecream>=2.1.5",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
